Apply unifier function done
This function takes two rows and an unifier, applies the unifier to the left row. Applying it to the right row means just copying it.
To achieve better performance, when a variable is replaced by a constant, only the first appearance of the variable changes value to the constant, the next appearances of the variable
still have a -index value. The unifier_a_b function handles this already. However. when a variable is replaced by another variable, things get a little tricky.

Let's say that x<-y, and y is a variable (x is always a variable). if row_a=[1,a,2,b] and row_b=[1,c,2,c], in the first substitution (a<-c), we do not need to do anything, but
in the next substitution (b<-c), we need to make b point to a, like this [1,0,2,-2]. For that I use a dictionary, and in the worst case scenario (y is a variable and it has substituted something before), we need an extra loop changing the references of x to the previous x. I think this gives better performance overall at the price of worse performance in the worst case scenario,
so results can vary greatly from one case (benchmark) to another. We could always do the the extra loop and change all references to x by the constant, that way we get a (i think maybe
worse) but more consistent performance. Will add this to discuss topics